get /synagogue api is done, added request api on frotn take all values from form in registration page

+++ backend/src/main/java/db/course_work/backend/repositories/SynagogueRepository.java
@@ -4,6 +4,9 @@ import db.course_work.backend.entities.Synagogue;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.List;
+
 @Repository
-public interface SynagogueRepository extends JpaRepository<Synagogue, Long> {
+public interface SynagogueRepository extends JpaRepository<Synagogue, Integer> {
+
 }
